refactor: move chartColors to constants folder
