+++ src/components/pool-detail/PriceChart.jsx
@@ -9,7 +9,7 @@ import {
    ReferenceLine
 } from "recharts"
 import { CustomPriceTooltip } from "./CustomPriceTooltip"
-import { CHART_COLORS } from "../../utils/chartColors"
+import { CHART_COLORS } from "../../constants/chartColors"
 import { formatCompactCurrency } from "../../utils/formatCompactCurrency"
 
 export function PriceChart({
