+++ src/components/pool-detail/TVLVolumeChart.jsx
@@ -11,7 +11,7 @@ import {
    ResponsiveContainer
 } from "recharts"
 import { CustomTooltip } from "./CustomTooltip"
-import { CHART_COLORS } from "../../utils/chartColors"
+import { CHART_COLORS } from "../../constants/chartColors"
 import { formatCompactCurrency } from "../../utils/formatCompactCurrency"
 
 export function TVLVolumeChart({ history }) {
